requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fasttext==0.9.2
filelock==3.13.1
Flask==2.3.2
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lightning-utilities==0.9.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nlpcube==0.3.1.2
numpy==1.25.0
packaging==23.2
psycopg2-binary==2.9.9
pybind11==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytorch-lightning==2.0.3
PyYAML==6.0
regex==2023.10.3
requests==2.31.0
rowordnet==1.1.0
safetensors==0.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.2.0
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
